Using short in PcmSoundPacket.sequence
git-svn-id: https://sovereigncomputing.net/svn/sneer/trunk@2952 ee3b4d21-14f5-0310-a1fa-c5d8b2925ec4

+++ sneer/bricks/src/snapps/listentome/speextuples/impl/SpeexTuplesImpl.java
@@ -68,6 +68,6 @@ class SpeexTuplesImpl implements SpeexTuples {
 	
 	protected void decode(SpeexPacket packet) {
 		for (byte[] frame : _decoder.decode(packet._frames))
-			_tupleSpace.acquire(new PcmSoundPacket(packet.publisher(), packet.publicationTime(), new ImmutableByteArray(frame, frame.length), 0));
+			_tupleSpace.acquire(new PcmSoundPacket(packet.publisher(), packet.publicationTime(), new ImmutableByteArray(frame, frame.length), (short)0));
 	}
 }
